Consolidate cloud-functions-chain into gcp/cloud-run/nodejs
- Move service-a/b/c from gcp/cloud-functions-chain/ into gcp/cloud-run/nodejs/
- Rename service-a → function-a to reflect it is a Cloud Function (gen2)
- Rename tracing.js → instrumentation.js in all services for consistency
- Update all require('./tracing') → require('./instrumentation')
- Remove hardcoded OTLP endpoints from .env.example and README files
- Add .env.example to function-a, service-b, service-c
- Update top-level README with chain deploy and local testing instructions
