--- gcp/cloud-functions-chain/.env.example
@@ -1,25 +1,0 @@
-# OpenTelemetry Configuration
-# Replace with your actual OTLP endpoint
-OTEL_EXPORTER_OTLP_ENDPOINT=YOUR_OTLP_ENDPOINT
-
-# Replace with your actual authentication credentials
-# Format: Authorization=Basic <base64-encoded-credentials>
-OTEL_EXPORTER_OTLP_HEADERS=YOUR_OTLP_HEADERS
-
-# Trace sampling configuration
-OTEL_TRACES_SAMPLER=always_on
-
-# Deployment environment
-OTEL_RESOURCE_ATTRIBUTES=deployment.environment=production
-
-# Log level
-OTEL_LOG_LEVEL=error
-
-# Service URLs (update these with your actual GCP URLs after deployment)
-SERVICE_B_URL=YOUR_SERVICE_B_URL
-SERVICE_C_URL=YOUR_SERVICE_C_URL
-
-# GCP Configuration
-GOOGLE_CLOUD_PROJECT=YOUR_PROJECT_ID
-FUNCTION_REGION=YOUR_REGION
-DEPLOYMENT_ENVIRONMENT=production
