+++ gcp/cloud-run/nodejs/express/.env.example
@@ -1,6 +1,6 @@
 # OpenTelemetry Configuration
 OTEL_SERVICE_NAME=express-cloud-run-demo
-OTEL_EXPORTER_OTLP_ENDPOINT=https://otlp.last9.io
+OTEL_EXPORTER_OTLP_ENDPOINT=<your-otlp-endpoint>
 OTEL_EXPORTER_OTLP_HEADERS=Authorization=Basic YOUR_BASE64_CREDENTIALS
 OTEL_TRACES_SAMPLER=always_on
