+++ gcp/cloud-run/nodejs/function-a/Dockerfile
@@ -2,17 +2,12 @@ FROM node:20-slim
 
 WORKDIR /app
 
-# Copy package files
 COPY package*.json ./
 
-# Install dependencies
 RUN npm install --production
 
-# Copy application files
 COPY . .
 
-# Expose port
 EXPOSE 8080
 
-# Start the application
 CMD ["npm", "start"]
