+++ gcp/cloud-run/nodejs/service-b/.env.example
@@ -1,0 +1,9 @@
+OTEL_EXPORTER_OTLP_ENDPOINT=<your-otlp-endpoint>
+OTEL_EXPORTER_OTLP_HEADERS="Authorization=Basic <your-last9-token>"
+OTEL_TRACES_SAMPLER=always_on
+OTEL_RESOURCE_ATTRIBUTES=deployment.environment=local
+OTEL_LOG_LEVEL=error
+OTEL_SERVICE_NAME=service-b
+
+# Service-to-service URL
+SERVICE_C_URL=<service-c-url>
