+++ gcp/cloud-run/nodejs/service-b/Dockerfile
@@ -2,17 +2,12 @@ FROM node:20-slim
 
 WORKDIR /app
 
-# Copy package files
 COPY package*.json ./
 
-# Install dependencies
 RUN npm install --production
 
-# Copy application files
 COPY . .
 
-# Expose port
 EXPOSE 8080
 
-# Start the Functions Framework
 CMD ["npm", "start"]
