+++ gcp/cloud-run/nodejs/service-c/Dockerfile
@@ -1,0 +1,13 @@
+FROM node:20-slim
+
+WORKDIR /app
+
+COPY package*.json ./
+
+RUN npm install --production
+
+COPY . .
+
+EXPOSE 8080
+
+CMD ["npm", "start"]
