+++ gcp/cloud-run/nodejs/service/.env.example
@@ -1,5 +1,5 @@
 # OpenTelemetry Configuration
-OTEL_EXPORTER_OTLP_ENDPOINT=https://otlp.last9.io
+OTEL_EXPORTER_OTLP_ENDPOINT=<your-otlp-endpoint>
 OTEL_EXPORTER_OTLP_HEADERS=Authorization=Basic <your-credentials>
 OTEL_SERVICE_NAME=cloud-run-service
